requirements.txt
translate>=3.6.1
pyyaml>=6.0
openai-whisper
ffmpeg-python>=0.2.0